fix: resolve React Fast Refresh ESLint warnings
Separate hooks and contexts from component files to improve Fast Refresh reliability.

Changes:
- Extract `buttonVariants` to `button-variants.ts`
- Extract `ChatContext` and `useChatContext` to `ChatContext.ts`
- Extract `ClientContext` and `useClient` to `ClientContext.ts`
- Update imports in consuming files

This fixes 4 `react-refresh/only-export-components` warnings.
