+++ src/providers/ChatContext.ts
@@ -1,0 +1,18 @@
+"use client";
+
+import { createContext, useContext } from "react";
+import type { useChat } from "@/app/hooks/useChat";
+
+export type ChatContextType = ReturnType<typeof useChat>;
+
+export const ChatContext = createContext<ChatContextType | undefined>(
+  undefined
+);
+
+export function useChatContext() {
+  const context = useContext(ChatContext);
+  if (context === undefined) {
+    throw new Error("useChatContext must be used within a ChatProvider");
+  }
+  return context;
+}
