+++ src/providers/ChatProvider.tsx
@@ -1,9 +1,10 @@
 "use client";
 
-import { ReactNode, createContext, useContext } from "react";
+import { ReactNode } from "react";
 import { Assistant } from "@langchain/langgraph-sdk";
 import { type StateType, useChat } from "@/app/hooks/useChat";
 import type { UseStreamThread } from "@langchain/langgraph-sdk/react";
+import { ChatContext } from "./ChatContext";
 
 interface ChatProviderProps {
   children: ReactNode;
@@ -22,16 +23,3 @@ export function ChatProvider({
   return <ChatContext.Provider value={chat}>{children}</ChatContext.Provider>;
 }
 
-export type ChatContextType = ReturnType<typeof useChat>;
-
-export const ChatContext = createContext<ChatContextType | undefined>(
-  undefined
-);
-
-export function useChatContext() {
-  const context = useContext(ChatContext);
-  if (context === undefined) {
-    throw new Error("useChatContext must be used within a ChatProvider");
-  }
-  return context;
-}
