+++ src/providers/ClientProvider.tsx
@@ -1,13 +1,8 @@
 "use client";
 
-import { createContext, useContext, useMemo, ReactNode } from "react";
+import { useMemo, ReactNode } from "react";
 import { Client } from "@langchain/langgraph-sdk";
-
-interface ClientContextValue {
-  client: Client;
-}
-
-const ClientContext = createContext<ClientContextValue | null>(null);
+import { ClientContext } from "./ClientContext";
 
 interface ClientProviderProps {
   children: ReactNode;
@@ -35,13 +30,4 @@ export function ClientProvider({
   return (
     <ClientContext.Provider value={value}>{children}</ClientContext.Provider>
   );
-}
-
-export function useClient(): Client {
-  const context = useContext(ClientContext);
-
-  if (!context) {
-    throw new Error("useClient must be used within a ClientProvider");
-  }
-  return context.client;
 }
